Extract useTasks hook
